refactor(batch): move InitialSegmentFootprint to its own file
The cross-segment gate grew to ~150 lines with the add-on build-out;
batch/mod.rs already splits comparable concerns into sub-files
(batch_structure, backward_references, indexed_tree). No behavior
change.
